+++ modules/azure-databricks-security-groups/variables.tf
@@ -9,7 +9,7 @@ variable "location" {
   description = "Location in which Databricks will be deployed"
 }
 
-variable "virtual_network_rg_name" {
+variable "vnet_resource_group_name" {
   type        = string
   description = "Name of resource group which contains the virtual network and subnets"
 }