+++ variables.tf
@@ -18,7 +18,7 @@ variable "vnet_id" {
   description = "ID of existing virtual network into which Databricks will be deployed"
 }
 
-variable "virtual_network_rg_name" {
+variable "vnet_resource_group_name" {
   type        = string
   description = "Name of resource group which contains the virtual network and subnets"
 }